+++ BarManagerApplication.Backend/src/BackendGestionaleBar.BusinessLayer/MapperProfiles/UserMapperProfile.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using BackendGestionaleBar.Authentication.Entities;
+using BackendGestionaleBar.Shared.Models;
 using BackendGestionaleBar.Shared.Requests;
 
 namespace BackendGestionaleBar.BusinessLayer.MapperProfiles;
@@ -8,6 +9,7 @@ internal class UserMapperProfile : Profile
 {
 	public UserMapperProfile()
 	{
+		CreateMap<ApplicationUser, User>();
 		CreateMap<RegisterUserRequest, ApplicationUser>();
 	}
 }
