--- BarManagerApplication.Core/src/BarApplication.APIClient/BarApplication.APIClient/Identity/Models/Requests/LoginRequest.cs
@@ -1,8 +1,0 @@
-﻿namespace BarApplication.APIClient.Identity.Models.Requests
-{
-    public class LoginRequest
-    {
-        public string UserName { get; set; }
-        public string Password { get; set; }
-    }
-}
