--- BarManagerApplication.Core/src/BarApplication.APIClient/BarApplication.APIClient/Identity/Models/Requests/RefreshTokenRequest.cs
@@ -1,8 +1,0 @@
-﻿namespace BarApplication.APIClient.Identity.Models.Requests
-{
-    public class RefreshTokenRequest
-    {
-        public string AccessToken { get; set; }
-        public string RefreshToken { get; set; }
-    }
-}
