--- BarManagerApplication.Core/src/BarApplication.APIClient/BarApplication.APIClient/Identity/Models/Requests/RegisterRequest.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace BarApplication.APIClient.Identity.Models.Requests
-{
-    public class RegisterRequest
-    {
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public DateTime BirthDate { get; set; }
-        public string PhoneNumber { get; set; }
-        public string Email { get; set; }
-        public string UserName { get; set; }
-        public string Password { get; set; }
-    }
-}
