--- BarManagerApplication.Core/src/BarApplication.APIClient/BarApplication.APIClient/Identity/Models/Requests/UpdatePasswordRequest.cs
@@ -1,8 +1,0 @@
-﻿namespace BarApplication.APIClient.Identity.Models.Requests
-{
-    public class UpdatePasswordRequest
-    {
-        public string Email { get; set; }
-        public string NewPassword { get; set; }
-    }
-}
