--- BarManagerApplication.Core/src/BarApplication.APIClient/BarApplication.APIClient/Identity/Models/Responses/AuthResponse.cs
@@ -1,8 +1,0 @@
-﻿namespace BarApplication.APIClient.Identity.Models.Responses
-{
-    public class AuthResponse
-    {
-        public string AccessToken { get; set; }
-        public string RefreshToken { get; set; }
-    }
-}
