--- BarManagerApplication.Core/src/BarApplication.APIClient/BarApplication.APIClient/Identity/Models/Responses/RegisterResponse.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace BarApplication.APIClient.Identity.Models.Responses
-{
-    public class RegisterResponse
-    {
-        public bool Succeeded { get; set; }
-        public IEnumerable<string> Errors { get; set; }
-    }
-}
