--- BarManagerApplication.Core/src/BarApplication.APIClient/BarApplication.APIClient/Identity/Models/User.cs
@@ -1,19 +1,0 @@
-﻿using System;
-
-namespace BarApplication.APIClient.Identity.Models
-{
-    public class User
-    {
-        public string FirstName { get; set; }
-
-        public string LastName { get; set; }
-
-        public DateTime BirthDate { get; set; }
-
-        public string Email { get; set; }
-
-        public string UserName { get; set; }
-
-        public string PhoneNumber { get; set; }
-    }
-}
